# packages/evidentia-ui/openapi.json
{
  "components": {
    "schemas": {
...
      "EfficiencyOpportunity": {
        "additionalProperties": false,
        "description": "A control that satisfies multiple framework requirements simultaneously.\n\nThese are high-value implementation targets — implementing one control\ncloses gaps across multiple frameworks.",
        "properties": {
          "control_id": {
            "description": "The NIST 800-53 control ID (canonical reference)",
            "title": "Control Id",
            "type": "string"
          },
          "control_title": {
            "description": "Human-readable control title",
            "title": "Control Title",
            "type": "string"
          },
          "framework_count": {
            "description": "Number of distinct frameworks satisfied",
            "title": "Framework Count",
            "type": "integer"
          },
          "frameworks_satisfied": {
            "description": "List of framework:control_id pairs this satisfies",
            "items": {
              "type": "string"
            },
            "title": "Frameworks Satisfied",
            "type": "array"
          },
          "implementation_effort": {
            "$ref": "#/components/schemas/ImplementationEffort",
            "description": "Estimated effort to implement"
          },
          "total_gaps_closed": {
            "description": "Total number of gap entries that would be closed",
            "title": "Total Gaps Closed",
            "type": "integer"
          },
          "value_score": {
            "description": "Efficiency value score = total_gaps_closed / effort_weight",
            "title": "Value Score",
            "type": "number"
          }
        },
        "required": [
          "control_id",
          "control_title",
          "frameworks_satisfied",
          "framework_count",
          "total_gaps_closed",
          "implementation_effort",
          "value_score"
        ],
        "title": "EfficiencyOpportunity",
        "type": "object"
      },
...
      "GapAnalysisReport-Output": {
        "additionalProperties": false,
        "description": "Complete gap analysis report.\n\nThe primary output of the gap analyzer. Contains all identified gaps,\nefficiency opportunities, and a prioritized remediation roadmap.",
        "properties": {
          "analyzed_at": {
            "format": "date-time",
            "title": "Analyzed At",
            "type": "string"
          },
          "coverage_percentage": {
            "description": "Percentage of required controls that are fully implemented",
            "title": "Coverage Percentage",
            "type": "number"
          },
          "critical_gaps": {
            "title": "Critical Gaps",
            "type": "integer"
          },
          "efficiency_opportunities": {
            "description": "Controls that satisfy 3+ framework requirements",
            "items": {
              "$ref": "#/components/schemas/EfficiencyOpportunity"
            },
            "title": "Efficiency Opportunities",
            "type": "array"
          },
          "evidentia_version": {
            "description": "Version of evidentia-core that produced this report",
            "title": "Evidentia Version",
            "type": "string"
          },
          "frameworks_analyzed": {
            "description": "Framework IDs that were analyzed",
            "items": {
              "type": "string"
            },
            "title": "Frameworks Analyzed",
            "type": "array"
          },
          "gaps": {
            "description": "All identified gaps, sorted by priority_score descending",
            "items": {
              "$ref": "#/components/schemas/ControlGap-Output"
            },
            "title": "Gaps",
            "type": "array"
          },
          "high_gaps": {
            "title": "High Gaps",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "informational_gaps": {
            "default": 0,
            "title": "Informational Gaps",
            "type": "integer"
          },
          "inventory_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Path to the inventory file used",
            "title": "Inventory Source"
          },
          "low_gaps": {
            "title": "Low Gaps",
            "type": "integer"
          },
          "medium_gaps": {
            "title": "Medium Gaps",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "organization": {
            "description": "Organization name from the control inventory",
            "title": "Organization",
            "type": "string"
          },
          "prioritized_roadmap": {
            "description": "Ordered list of gap IDs by descending priority_score",
            "items": {
              "type": "string"
            },
            "title": "Prioritized Roadmap",
            "type": "array"
          },
          "total_controls_in_inventory": {
            "description": "Total controls in the organization's inventory",
            "title": "Total Controls In Inventory",
            "type": "integer"
          },
          "total_controls_required": {
            "description": "Total unique controls required across all analyzed frameworks",
            "title": "Total Controls Required",
            "type": "integer"
          },
          "total_gaps": {
            "title": "Total Gaps",
            "type": "integer"
          }
        },
        "required": [
          "organization",
          "frameworks_analyzed",
          "total_controls_required",
          "total_controls_in_inventory",
          "total_gaps",
          "critical_gaps",
          "high_gaps",
          "medium_gaps",
          "low_gaps",
          "coverage_percentage",
          "gaps"
        ],
        "title": "GapAnalysisReport",
        "type": "object"
      },
      "GapAnalyzeRequest": {
        "additionalProperties": false,
        "description": "Body of `POST /api/gap/analyze`.\n\nEither ``inventory_path`` (server-side file) or ``inventory_content``\n(inline YAML/JSON uploaded from the browser) must be provided.",
        "properties": {
          "frameworks": {
            "description": "One or more framework IDs to analyze against.",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Frameworks",
            "type": "array"
          },
          "inventory_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Inline inventory YAML/JSON (browser upload).",
            "title": "Inventory Content"
          },
          "inventory_format": {
            "default": "yaml",
            "description": "One of: yaml, json, csv.",
            "title": "Inventory Format",
            "type": "string"
          },
          "inventory_path": {
            "anyOf": [
              {
                "format": "path",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Absolute path to inventory file (server-side).",
            "title": "Inventory Path"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override inventory.organization.",
            "title": "Organization"
          },
          "system_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Override inventory.system_name.",
            "title": "System Name"
          }
        },
        "required": [
          "frameworks"
        ],
        "title": "GapAnalyzeRequest",
        "type": "object"
      },
      "GapDiff": {
        "additionalProperties": false,
        "description": "A full gap-diff document comparing two GapAnalysisReport snapshots.",
        "properties": {
          "base_inventory_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Inventory Source"
          },
          "base_organization": {
            "description": "Organization name from base report",
            "title": "Base Organization",
            "type": "string"
          },
          "entries": {
            "description": "Every gap key present in base or head (or both)",
            "items": {
              "$ref": "#/components/schemas/GapDiffEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "frameworks_analyzed": {
            "description": "Union of frameworks analyzed in either report",
            "items": {
              "type": "string"
            },
            "title": "Frameworks Analyzed",
            "type": "array"
          },
          "generated_at": {
            "description": "ISO 8601 timestamp when this diff was generated",
            "title": "Generated At",
            "type": "string"
          },
          "head_inventory_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Inventory Source"
          },
          "head_organization": {
            "description": "Organization name from head report",
            "title": "Head Organization",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/GapDiffSummary",
            "description": "Per-status aggregate counts"
          }
        },
        "required": [
          "base_organization",
          "head_organization",
          "frameworks_analyzed"
        ],
        "title": "GapDiff",
        "type": "object"
      },
...
      "GapDiffRequest": {
        "additionalProperties": false,
        "description": "Body of `POST /api/gap/diff`.",
        "properties": {
          "base_key": {
            "description": "gap_store SHA-16 key of the base report.",
            "title": "Base Key",
            "type": "string"
          },
          "head_key": {
            "description": "gap_store SHA-16 key of the head report.",
            "title": "Head Key",
            "type": "string"
          }
        },
        "required": [
          "base_key",
          "head_key"
        ],
        "title": "GapDiffRequest",
        "type": "object"
      },
      "GapDiffSummary": {
        "additionalProperties": false,
        "description": "Aggregate counts per diff status — for quick CI decision-making.",
        "properties": {
          "closed": {
            "default": 0,
            "description": "Number of gaps resolved",
            "title": "Closed",
            "type": "integer"
          },
          "opened": {
            "default": 0,
            "description": "Number of new gaps (regressions)",
            "title": "Opened",
            "type": "integer"
          },
          "severity_decreased": {
            "default": 0,
            "title": "Severity Decreased",
            "type": "integer"
          },
          "severity_increased": {
            "default": 0,
            "title": "Severity Increased",
            "type": "integer"
          },
          "unchanged": {
            "default": 0,
            "title": "Unchanged",
            "type": "integer"
          }
        },
        "title": "GapDiffSummary",
        "type": "object"
      },
      "GapExportRequest": {
        "additionalProperties": false,
        "description": "Body of `POST /api/gap/export`.\n\nCarries a full :class:`GapAnalysisReport` (the object the UI already\nholds after `POST /api/gap/analyze`) plus a target ``format``. The\nendpoint reuses the CLI's :func:`evidentia_core.gap_analyzer.export_report`\nemitters and streams the serialized artifact back as a download — no\nre-implementation of any format on the API side.\n\n``report_key`` is accepted as an alternative source so a caller can\nexport a report already persisted in the gap store without re-sending\nthe whole body. Exactly one of ``report`` / ``report_key`` must be set.",
        "properties": {
          "format": {
            "default": "json",
            "description": "One of: json, csv, markdown, oscal-ar, sarif, ocsf, ocsf-detection, cyclonedx-vex. OCSF formats (ocsf, ocsf-detection) require the server's [ocsf] extra; cyclonedx-vex / oscal-ar are always available.",
            "title": "Format",
            "type": "string"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GapAnalysisReport-Input"
              },
              {
                "type": "null"
              }
            ],
            "description": "Inline report to export (as returned by /api/gap/analyze)."
          },
          "report_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "gap_store SHA-16 key of a saved report to export instead.",
            "title": "Report Key"
          }
        },
        "title": "GapExportRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImplementationEffort": {
        "description": "Estimated effort to remediate a gap.",
        "enum": [
          "low",
          "medium",
          "high",
          "very_high"
        ],
        "title": "ImplementationEffort",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "REST API for the Evidentia GRC tool. All endpoints mirror CLI capabilities. Binds to 127.0.0.1 by default for localhost web-UI use.",
    "title": "Evidentia API",
    "version": "0.10.7"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/gap/analyze": {
      "post": {
        "description": "Run gap analysis and persist the result to the gap store.\n\nReturns the freshly-computed report. The persisted JSON lives under\nthe user-dir gap_store and is referenced by the ``GET /api/gap/reports``\nendpoint after this call completes.",
        "operationId": "analyze_api_gap_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GapAnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapAnalysisReport-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze",
        "tags": [
          "gaps"
        ]
      }
    },
    "/api/gap/diff": {
      "post": {
        "description": "Compute a diff between two saved gap reports.\n\nReferenced by the /gap/diff page: user picks base + head from the\nreports list, clicks \"Compare\", and the result drives the summary +\nper-entry table.",
        "operationId": "diff_api_gap_diff_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GapDiffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapDiff"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff",
        "tags": [
          "gaps"
        ]
      }
    },
    "/api/gap/export": {
      "post": {
        "description": "Export a gap report in an engine-supported format as a download.\n\nReuses :func:`evidentia_core.gap_analyzer.export_report` — the exact\nemitters behind the CLI's ``evidentia gap analyze --format ...`` — so\nthe API never re-implements a serialization. The artifact is returned\nwith a ``Content-Disposition: attachment`` header so the browser saves\nit directly.\n\nThe report source is either an inline ``report`` body (what the GUI\nholds after running an analysis) or a ``report_key`` referencing a\nsaved gap-store report. Exactly one must be supplied.",
        "operationId": "export_api_gap_export_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GapExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export",
        "tags": [
          "gaps"
        ]
      }
    },
...
    "/api/gap/reports/{key}": {
      "get": {
        "description": "Load a saved gap report by its storage key.",
        "operationId": "get_report_api_gap_reports__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapAnalysisReport-Output"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report",
        "tags": [
          "gaps"
        ]
      }
    },
...
  }
}